[feature] extract ResumeDbService interface
